+++ src/WriteSeq32.cpp
@@ -58,6 +58,9 @@ struct WriteSeq32 : Module {
 		NUM_LIGHTS
 	};
 
+	// Constants
+	// none
+	
 	// Need to save, no reset
 	int panelTheme;
 	
@@ -69,7 +72,6 @@ struct WriteSeq32 : Module {
 	float cv[4][32];
 	int gates[4][32];
 	bool resetOnRun;
-	unsigned long editingGate;// 0 when no edit gate, downward step counter timer when edit gate
 
 	// No need to save, with reset
 	long clockIgnoreOnReset;
@@ -77,6 +79,8 @@ struct WriteSeq32 : Module {
 	int gateCPbuffer[32];// copy paste buffer for gates
 	long infoCopyPaste;// 0 when no info, positive downward step counter timer when copy, negative upward when paste
 	int pendingPaste;// 0 = nothing to paste, 1 = paste on clk, 2 = paste on seq, destination channel in next msbits
+	unsigned long editingGate;// 0 when no edit gate, downward step counter timer when edit gate
+	int stepRotates;
 
 	// No need to save, no reset
 	RefreshCounter refresh;	
@@ -155,6 +159,7 @@ struct WriteSeq32 : Module {
 		infoCopyPaste = 0l;
 		pendingPaste = 0;
 		editingGate = 0ul;
+		stepRotates = 0;
 	}
 
 	
@@ -370,14 +375,38 @@ struct WriteSeq32 : Module {
 				delta = +1;
 			}
 			if (delta != 0 && canEdit) {		
-				if (indexChannel == 3)
-					indexStepStage = moveIndex(indexStepStage, indexStepStage + delta, numSteps);
-				else 
-					indexStep = moveIndex(indexStep, indexStep + delta, numSteps);
-				// Editing gate
-				int index = (indexChannel == 3 ? indexStepStage : indexStep);
-				editingGate = (unsigned long) (gateTime * args.sampleRate / RefreshCounter::displayRefreshStepSkips);
-				editingGateCV = cv[indexChannel][index];
+				if (stepRotates == 0) {
+					if (indexChannel == 3)
+						indexStepStage = moveIndex(indexStepStage, indexStepStage + delta, numSteps);
+					else 
+						indexStep = moveIndex(indexStep, indexStep + delta, numSteps);
+					// Editing gate
+					int index = (indexChannel == 3 ? indexStepStage : indexStep);
+					editingGate = (unsigned long) (gateTime * args.sampleRate / RefreshCounter::displayRefreshStepSkips);
+					editingGateCV = cv[indexChannel][index];
+				}
+				else {// rotate mode
+					float rotCV;
+					int rotGate;
+					int iStart = 0;
+					int iEnd = numSteps - 1;
+					int iRot = iStart;
+					int iDelta = 1;
+					if (delta == 1) {
+						iRot = iEnd;
+						iDelta = -1;
+					}
+					rotCV = cv[indexChannel][iRot];
+					rotGate = gates[indexChannel][iRot];
+					for ( ; ; iRot += iDelta) {
+						if (iDelta == 1 && iRot >= iEnd) break;
+						if (iDelta == -1 && iRot <= iStart) break;
+						cv[indexChannel][iRot] = cv[indexChannel][iRot + iDelta];
+						gates[indexChannel][iRot] = gates[indexChannel][iRot + iDelta];
+					}
+					cv[indexChannel][iRot] = rotCV;
+					gates[indexChannel][iRot] = rotGate;
+				}
 			}
 			
 			// Window buttons
@@ -609,6 +638,32 @@ struct WriteSeq32Widget : ModuleWidget {
 			module->resetOnRun = !module->resetOnRun;
 		}
 	};
+	
+	struct ArrowModeItem : MenuItem {
+		int* stepRotatesSrc;
+
+		struct ArrowModeSubItem : MenuItem {
+			int* stepRotatesSrc;
+			void onAction(const event::Action &e) override {
+				*stepRotatesSrc ^= 0x1;
+			}
+		};
+
+		Menu *createChildMenu() override {
+			Menu *menu = new Menu;
+			
+			ArrowModeSubItem *step0Item = createMenuItem<ArrowModeSubItem>("Step", CHECKMARK(*stepRotatesSrc == 0));
+			step0Item->stepRotatesSrc = stepRotatesSrc;
+			menu->addChild(step0Item);
+
+			ArrowModeSubItem *step1Item = createMenuItem<ArrowModeSubItem>("Rotate", CHECKMARK(*stepRotatesSrc != 0));
+			step1Item->stepRotatesSrc = stepRotatesSrc;
+			menu->addChild(step1Item);
+
+			return menu;
+		}
+	};	
+
 	void appendContextMenu(Menu *menu) override {
 		MenuLabel *spacerLabel = new MenuLabel();
 		menu->addChild(spacerLabel);
@@ -632,6 +687,10 @@ struct WriteSeq32Widget : ModuleWidget {
 		settingsLabel->text = "Settings";
 		menu->addChild(settingsLabel);
 		
+		ArrowModeItem *arrowItem = createMenuItem<ArrowModeItem>("Arrow controls", RIGHT_ARROW);
+		arrowItem->stepRotatesSrc = &(module->stepRotates);
+		menu->addChild(arrowItem);
+
 		ResetOnRunItem *rorItem = createMenuItem<ResetOnRunItem>("Reset on run", CHECKMARK(module->resetOnRun));
 		rorItem->module = module;
 		menu->addChild(rorItem);